Create user class, dom update and data manipulation files, move functions from index.js.
